Remove fee related stuff
